+++ portal/views/api.py
@@ -1,7 +1,7 @@
 import datetime
 import uuid
 
-from common.models import Student, Teacher
+from common.models import Student, Teacher, Class
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.http import HttpRequest, HttpResponse
@@ -21,9 +21,7 @@ THREE_YEARS_IN_DAYS = 1095
 @login_required(login_url=reverse_lazy("administration_login"))
 def registered_users(request, year, month, day):
     try:
-        nbr_reg = User.objects.filter(
-            date_joined__startswith=datetime.date(int(year), int(month), int(day))
-        ).count()
+        nbr_reg = User.objects.filter(date_joined__startswith=datetime.date(int(year), int(month), int(day))).count()
         return Response(nbr_reg)
     except ValueError:
         return HttpResponse(status=404)
@@ -33,9 +31,7 @@ def registered_users(request, year, month, day):
 @login_required(login_url=reverse_lazy("administration_login"))
 def last_connected_since(request, year, month, day):
     try:
-        nbr_active_users = User.objects.filter(
-            last_login__gte=datetime.date(int(year), int(month), int(day))
-        ).count()
+        nbr_active_users = User.objects.filter(last_login__gte=datetime.date(int(year), int(month), int(day))).count()
         return Response(nbr_active_users)
     except ValueError:
         return HttpResponse(status=404)
@@ -47,9 +43,7 @@ def number_users_per_country(request, country):
     try:
         nbr_reg = (
             Teacher.objects.filter(school__country__exact=country).count()
-            + Student.objects.filter(
-                class_field__teacher__school__country__exact=country
-            ).count()
+            + Student.objects.filter(class_field__teacher__school__country__exact=country).count()
         )
         return Response(nbr_reg)
     except ValueError:
@@ -75,13 +69,60 @@ class IsAdminOrGoogleAppEngine(permissions.IsAdminUser):
         return IS_CLOUD_SCHEDULER_FUNCTION(request) or is_admin
 
 
-def anonymise(user):
+def __anonymise_user(user):
+    # the actual user anonymisation
     user.username = uuid.uuid4().hex
     user.first_name = "Deleted"
     user.last_name = "User"
     user.email = ""
     user.is_active = False
     user.save()
+
+
+def anonymise(user):
+    """Anonymise user. If admin teacher, pass the admin role to another teacher (if exists).
+    If the only teacher, anonymise the school.
+    """
+    is_admin = False
+    teacher = None
+    teacher_set = Teacher.objects.filter(new_user=user)
+    if teacher_set:
+        is_admin = teacher_set[0].is_admin
+        school = teacher_set[0].school
+        teacher = teacher_set[0]
+
+    __anonymise_user(user)
+
+    # if teacher, clean up classes and anonymise students
+    if teacher:
+        # delete classes
+        classes = Class.objects.filter(teacher=teacher)
+        for klass in classes:
+            # anonymise the students
+            students = Student.objects.filter(class_field=klass)
+            for student in students:
+                __anonymise_user(student.new_user)
+            klass.delete()  # TODO: update to anonymise later
+
+    # if user is admin and the school does not have another admin, appoint another teacher as admin
+    if is_admin:
+        teachers = Teacher.objects.filter(school=school).order_by("new_user__last_name", "new_user__first_name")
+        if not teachers:
+            # no other teacher, scramble the school name
+            school.name = uuid.uuid4().hex
+            school.save()
+            return
+
+        admin_exists = False
+        for teacher in teachers:
+            if teacher.is_admin:
+                admin_exists = True
+                break
+
+        # if no admin, appoint the first teacher as admin
+        if not admin_exists:
+            teachers[0].is_admin = True
+            teachers[0].save()
 
 
 class InactiveUsersView(generics.ListAPIView):
@@ -93,14 +134,10 @@ class InactiveUsersView(generics.ListAPIView):
     """
 
     queryset = User.objects.filter(is_active=True) & (
-        User.objects.filter(
-            last_login__lte=timezone.now()
-            - timezone.timedelta(days=THREE_YEARS_IN_DAYS)
-        )
+        User.objects.filter(last_login__lte=timezone.now() - timezone.timedelta(days=THREE_YEARS_IN_DAYS))
         | User.objects.filter(
             last_login__isnull=True,
-            date_joined__lte=timezone.now()
-            - timezone.timedelta(days=THREE_YEARS_IN_DAYS),
+            date_joined__lte=timezone.now() - timezone.timedelta(days=THREE_YEARS_IN_DAYS),
         )
     )
     authentication_classes = (SessionAuthentication,)